conn: move conn code to a separate file
